requirements.txt
fastapi==0.103.1
python-multipart==0.0.6
python-dotenv==1.0.0
uvicorn==0.23.2
Pillow==10.0.1
requests==2.31.0